// frontend/src/api/generated/models/OperationalExpenseTemplateResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Rental API
 * API for rental management
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { RoomResponse } from './RoomResponse';
import {
    RoomResponseFromJSON,
    RoomResponseFromJSONTyped,
    RoomResponseToJSON,
} from './RoomResponse';
import type { ExpenseCategory } from './ExpenseCategory';
import {
    ExpenseCategoryFromJSON,
    ExpenseCategoryFromJSONTyped,
    ExpenseCategoryToJSON,
} from './ExpenseCategory';
import type { ApartmentResponse } from './ApartmentResponse';
import {
    ApartmentResponseFromJSON,
    ApartmentResponseFromJSONTyped,
    ApartmentResponseToJSON,
} from './ApartmentResponse';

/**
 * 
 * @export
 * @interface OperationalExpenseTemplateResponse
 */
export interface OperationalExpenseTemplateResponse {
    /**
     * 
     * @type {number}
     * @memberof OperationalExpenseTemplateResponse
     */
    id: number;
    /**
     * 
     * @type {ApartmentResponse}
     * @memberof OperationalExpenseTemplateResponse
     */
    apartment?: ApartmentResponse | null;
    /**
     * 
     * @type {RoomResponse}
     * @memberof OperationalExpenseTemplateResponse
     */
    room?: RoomResponse | null;
    /**
     * 
     * @type {number}
     * @memberof OperationalExpenseTemplateResponse
     */
    amount: number;
    /**
     * 
     * @type {ExpenseCategory}
     * @memberof OperationalExpenseTemplateResponse
     */
    category: ExpenseCategory;
    /**
     * 
     * @type {number}
     * @memberof OperationalExpenseTemplateResponse
     */
    dayOfMonth: number;
    /**
     * 
     * @type {boolean}
     * @memberof OperationalExpenseTemplateResponse
     */
    active: boolean;
}



/**
 * Check if a given object implements the OperationalExpenseTemplateResponse interface.
 */
export function instanceOfOperationalExpenseTemplateResponse(value: object): value is OperationalExpenseTemplateResponse {
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('amount' in value) || value['amount'] === undefined) return false;
    if (!('category' in value) || value['category'] === undefined) return false;
    if (!('dayOfMonth' in value) || value['dayOfMonth'] === undefined) return false;
    if (!('active' in value) || value['active'] === undefined) return false;
    return true;
}

export function OperationalExpenseTemplateResponseFromJSON(json: any): OperationalExpenseTemplateResponse {
    return OperationalExpenseTemplateResponseFromJSONTyped(json, false);
}

export function OperationalExpenseTemplateResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): OperationalExpenseTemplateResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'],
        'apartment': json['apartment'] == null ? undefined : ApartmentResponseFromJSON(json['apartment']),
        'room': json['room'] == null ? undefined : RoomResponseFromJSON(json['room']),
        'amount': json['amount'],
        'category': ExpenseCategoryFromJSON(json['category']),
        'dayOfMonth': json['dayOfMonth'],
        'active': json['active'],
    };
}

export function OperationalExpenseTemplateResponseToJSON(value?: OperationalExpenseTemplateResponse | null): any {
    if (value == null) {
        return value;
    }
    return {
        
        'id': value['id'],
        'apartment': ApartmentResponseToJSON(value['apartment']),
        'room': RoomResponseToJSON(value['room']),
        'amount': value['amount'],
        'category': ExpenseCategoryToJSON(value['category']),
        'dayOfMonth': value['dayOfMonth'],
        'active': value['active'],
    };
}